Move: MessageFactory -> Messages\Factory renamed moved Peer classes to correct folder
