+++ examples/listener.client.php
@@ -15,15 +15,16 @@ $host = $peerFactory->getAddress('127.0.0.1');
 $local = $peerFactory->getAddress('192.168.192.39', 32301);
 
 $peer = $peerFactory->getPeer();
-$peer->on('ready', function (Peer $peer) use ($factory) {
+$peer->on('ready', function (Peer $peer) use ($loop) {
     echo "connected\n";
     $peer->getaddr();
-    $peer->on('addr', function (Addr $addr) {
+    $peer->on('addr', function (Peer $peer, Addr $addr) use ($loop) {
         echo "Nodes: \n";
         foreach ($addr->getAddresses() as $address)
         {
             echo $address->getIp() . "\n";
         }
+        $loop->stop();
     });
 });
