+++ tests/Messages/FilterAddTest.php
@@ -4,7 +4,7 @@ namespace BitWasp\Bitcoin\Tests\Networking\Messages;
 
 use BitWasp\Bitcoin\Bitcoin;
 use BitWasp\Bitcoin\Crypto\Random\Random;
-use BitWasp\Bitcoin\Networking\MessageFactory;
+use BitWasp\Bitcoin\Networking\Messages\Factory;
 use BitWasp\Bitcoin\Tests\Networking\AbstractTestCase;
 use BitWasp\Buffertools\Buffer;
 use BitWasp\Buffertools\Parser;
@@ -15,7 +15,7 @@ class FilterAddTest extends AbstractTestCase
     {
         $data = new Buffer('aaaa');
 
-        $factory = new MessageFactory(Bitcoin::getDefaultNetwork(), new Random());
+        $factory = new Factory(Bitcoin::getDefaultNetwork(), new Random());
         $filteradd = $factory->filteradd($data);
 
         $serialized = $filteradd->getNetworkMessage()->getBuffer();
