+++ tests/Messages/FilterClearTest.php
@@ -4,7 +4,7 @@ namespace BitWasp\Bitcoin\Tests\Networking\Messages;
 
 use BitWasp\Bitcoin\Bitcoin;
 use BitWasp\Bitcoin\Crypto\Random\Random;
-use BitWasp\Bitcoin\Networking\MessageFactory;
+use BitWasp\Bitcoin\Networking\Messages\Factory;
 use BitWasp\Bitcoin\Tests\Networking\AbstractTestCase;
 use BitWasp\Buffertools\Parser;
 
@@ -13,7 +13,7 @@ class FilterClearTest extends AbstractTestCase
     public function testNetworkSerializable()
     {
 
-        $factory = new MessageFactory(Bitcoin::getDefaultNetwork(), new Random());
+        $factory = new Factory(Bitcoin::getDefaultNetwork(), new Random());
         $filterclear = $factory->filterclear();
 
         $serialized = $filterclear->getNetworkMessage()->getBuffer();
