+++ tests/Messages/GetAddrTest.php
@@ -4,7 +4,7 @@ namespace BitWasp\Bitcoin\Tests\Networking\Messages;
 
 use BitWasp\Bitcoin\Bitcoin;
 use BitWasp\Bitcoin\Crypto\Random\Random;
-use BitWasp\Bitcoin\Networking\MessageFactory;
+use BitWasp\Bitcoin\Networking\Messages\Factory;
 use BitWasp\Bitcoin\Networking\Serializer\NetworkMessageSerializer;
 use BitWasp\Buffertools\Buffer;
 use BitWasp\Bitcoin\Networking\Messages\GetAddr;
@@ -14,7 +14,7 @@ class GetAddrTest extends AbstractTestCase
 {
     public function testGetAddr()
     {
-        $factory = new MessageFactory(Bitcoin::getDefaultNetwork(), new Random());
+        $factory = new Factory(Bitcoin::getDefaultNetwork(), new Random());
         $getaddr = $factory->getaddr();
         $this->assertSame('getaddr', $getaddr->getNetworkCommand());
         $this->assertEquals(new Buffer(), $getaddr->getBuffer());
